Updates for renamed PVR_ERROR_UNKNOWN.
Pass all required fields across the interface (complete fakery for the EPG_TAG.iUniqueBroadcastId though).

Turn on support for recordings so they are displayed.

Trigger update of recordings when Addon is updated. Will need to hook in some other updates until the MythProtocol is plugged in and that can be done on the fly.